+++ test/sharedobject.test.ts
@@ -1095,8 +1095,8 @@ describe('SharedObject Pattern', () => {
       port = await getAvailablePort();
       descriptor = {
         transport: {
-          server: `127.0.0.1:${port}`,
-          client: `127.0.0.1:${port}`,
+          server: { host: '127.0.0.1', port },
+          client: { host: '127.0.0.1', port },
         },
         endpoints: [
           {